+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
-## v0.0.1
+## v0.1.1
 
-- Initial release
+-   Fixed default value for the prompt so that it won't appear empty if not 
+    supplied in the user config JSON.
+-   Fixed commands that set the configuration values to use the correct type.
+
+## v0.1.0
+
+-   Initial release